Se modifica la propieda de review en user y en demas casos
